Add support for unified authentication with OpenShift
Kubernetes as of 1.6 supports well factored authentication and
authorization plugins that allow a client to perform remote
authentication and authorization in a simple fashion. This upgrades the
oauth_proxy to support that as a fallback.

Remote authorization is opt in via the --openshift-delegate flag - if
requested the provided kubeconfig files or in-cluster config will be
checked, and if the user provides a bearer token or client cert that the
proxy recognizes the user will be allowed through if they match the
--openshift-sar flag reviews, or match the new --openshift-resources map
of paths to resource attributes.

This commit removes providers we don't intend to support and cleans up
some of the initialization code to allow additional flags to be passed.