Allow adding the system roots in addition to a specified ca file
This is to allow the openshift provider to be backwards compatible
when it sets the default, which happens every time it does not
have an explicit configuration. Otherwise system roots can never
be automatically queried.

Users are advise to explicitly set the openshift-ca parameter if
they do not want to allow system roots to be trusted.

Signed-off-by: Simo Sorce <simo@redhat.com>