Refine passing OAuth token from client upstream
The OAuth token present in the `Authorization` header is not passed to
upstream if delegating authentication/authorization to OpenShift using
`--openshift-delegate-urls`.

This adds new configuration parameter `--pass-user-bearer-token` that
enables or disables passing the received token from the client to the
upstream service.

This would extract the token from the `Authorization` header and pass it
to upstream via the usual `X-Forwarded-Access-Token` header.