openshift provider: refetch oauth endpoints on login
Adds refetching of oauth endpoints and endpoints based on them
when a user tries to do a login. This happens unless these are
hardcoded by user config when the binary is started.

The provider will try to refetch the information also in the cases
where the fetching during login originally failed, but the endpoint
information is required for whatever different purpose.